Moves ServiceProvider into the module.
